+++ src/Pim/Bundle/CatalogBundle/Validator/Constraints/UniqueValueValidator.php
@@ -18,9 +18,7 @@ use Symfony\Component\Form\Form;
  */
 class UniqueValueValidator extends ConstraintValidator
 {
-    /**
-     * @var ProductManager
-     */
+    /** @var ProductManager */
     protected $productManager;
 
     /**
@@ -34,26 +32,29 @@ class UniqueValueValidator extends ConstraintValidator
     }
 
     /**
-     * Constraint is applied on ProductValue data property.
-     * That's why we use the current property path to guess the code
-     * of the attribute to which the data belongs to.
-     * @param object     $rawValue
-     * @param Constraint $constraint
+     * Due to constraint guesser, the constraint is applied on :
+     * - ProductValueInterface data when applied through form
+     * - ProductValueInterface when applied directly through validator
+     *
+     * The constraint guesser should be re-worked in a future version to avoid such behavior
+     *
+     * @param ProductValueInterface|string $data
+     * @param Constraint                   $constraint
      *
      * @see Pim\Bundle\CatalogBundle\Validator\ConstraintGuesser\UniqueValueGuesser
      */
-    public function validate($rawValue, Constraint $constraint)
+    public function validate($data, Constraint $constraint)
     {
-        if (empty($rawValue)) {
+        if (empty($data)) {
             return;
         }
 
         $productValue = null;
-        $root = $this->context->getRoot();
-        if ($root instanceof Form) {
+        if (is_string($data)) {
             $productValue = $this->getProductValueFromForm();
+        } elseif (is_object($data) && $data instanceof ProductValueInterface) {
+            $productValue = $data;
         } else {
-            // TODO not supported through direct validation, how to get the value ?
             return;
         }
 
@@ -69,6 +70,11 @@ class UniqueValueValidator extends ConstraintValidator
      */
     protected function getProductValueFromForm()
     {
+        $root = $this->context->getRoot();
+        if (!$root instanceof Form) {
+            return;
+        }
+
         preg_match(
             '/children\[values\].children\[(\w+)\].children\[\w+\].data/',
             $this->context->getPropertyPath(),